requirements.txt
asgiref==3.8.1
blinker==1.8.2
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.1
djangorestframework==3.15.2
Flask==3.0.3
fonttools==4.53.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
plotly==5.24.1
PyAudio==0.2.14
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.4
pyproj==3.6.1
pyshp==2.3.1
python-dateutil==2.9.0.post0
requests==2.32.3
retrying==1.3.4
scapy==2.5.0
scipy==1.14.1
setuptools==75.2.0
shapely==2.0.6
six==1.16.0
sounddevice==0.5.0
SpeechRecognition==3.10.4
sqlparse==0.5.1
style==1.1.0
tenacity==9.0.0
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2

PyPDF2~=3.0.1
DateTime~=5.5